# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the Showcaster AI affiliate video generation SaaS. Handles user authentication, video generation jobs, and image uploads.",
        "title": "Showcaster API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Showcaster Support"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/upload/image": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uploads a JPEG or PNG image to Cloudinary and returns the public HTTPS URL.\nUse the returned URL as `modelImageUrl` or `productImageUrl` when submitting a job.\n\n**Constraints:**\n- File must be sent in the `file` multipart form field.\n- Accepted MIME types: `image/jpeg`, `image/png` (detected from file content, not the header).\n- Maximum file size: 10 MB.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Upload a model or product image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file (JPEG or PNG, max 10 MB)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upload successful — use the URL in job submission",
                        "schema": {
                            "$ref": "#/definitions/dto.UploadImageResponse"
                        }
                    },
                    "400": {
                        "description": "Missing `file` field in form data",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid JWT",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "File exceeds the 10 MB size limit",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "415": {
                        "description": "Unsupported file type (only JPEG and PNG accepted)",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Cloudinary upload failed",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "human-readable error message"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.UploadImageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.UploadImageResponseData"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.UploadImageResponseData": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string",
                    "example": "https://res.cloudinary.com/demo/image/upload/v1716000000/showcaster/images/abc123.jpg"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT token as: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}